Fix SSID retrevial fails for systems with non-english locales
Line 30 greps for a line starting with "yes" (the "active" field) but, in e.g. a French system, these fields will have "oui" and "non" instead of "yes" and "no".

Use the "in-use" field, which outputs an asterisk "*" for the in-use network and a space otherwise. grep for a line starting with "*".